Added "Levels" to Core Objects
Forgot to add a value which keeps track on when a certain Kanji/Vocabulary should be unlocked. To do this, each Kanji and Vocabulary will have a Level associated with them. Based on the user level at the current time of runtime, certain Kanji/Vocabulary will be unlocked.

Vocabulary will have an additional requirement - which is to have a high enough profeciency with the Kanji used in the Vocabulary. More on this when implementing "Grade.cs".